--- agent/agent.py
@@ -1,6 +1,0 @@
-class Agent():
-  def __init__(self):
-    pass
-
-  def act(self, observation):
-    return {}
