--- agent/baseline_agent.py
@@ -1,5 +1,0 @@
-
-from nmmo import Agent
-
-class BaselineAgent(Agent):
-  pass
