+++ environment.yml
@@ -1,0 +1,154 @@
+name: nmmo
+channels:
+  - conda-forge
+dependencies:
+  - bzip2=1.0.8=h3422bc3_4
+  - c-ares=1.18.1=h3422bc3_0
+  - ca-certificates=2022.12.7=h4653dfc_0
+  - grpcio=1.52.1=py39h23fbdae_1
+  - libabseil=20230125.0=cxx17_hb7217d7_1
+  - libcxx=15.0.7=h75e25f2_0
+  - libffi=3.4.2=h3422bc3_5
+  - libgrpc=1.52.1=he98ff75_1
+  - libprotobuf=3.21.12=hb5ab8b9_0
+  - libsqlite=3.40.0=h76d750c_0
+  - libzlib=1.2.13=h03a7124_4
+  - ncurses=6.3=h07bb92c_1
+  - openssl=3.0.8=h03a7124_0
+  - pip=23.0.1=pyhd8ed1ab_0
+  - python=3.9.16=hea58f1e_0_cpython
+  - python_abi=3.9=3_cp39
+  - re2=2023.02.02=hb7217d7_0
+  - readline=8.1.2=h46ed386_0
+  - setuptools=67.4.0=pyhd8ed1ab_0
+  - tk=8.6.12=he1e0b03_0
+  - tzdata=2022g=h191b570_0
+  - wheel=0.38.4=pyhd8ed1ab_0
+  - xz=5.2.6=h57fd34a_0
+  - zlib=1.2.13=h03a7124_4
+  - pip:
+    - absl-py==1.4.0
+    - aiosignal==1.3.1
+    - appdirs==1.4.4
+    - astroid==2.14.2
+    - attrs==22.2.0
+    - autobahn==19.3.3
+    - autoflake==2.0.2
+    - automat==22.10.0
+    - cachetools==5.3.0
+    - certifi==2022.12.7
+    - cffi==1.15.1
+    - charset-normalizer==3.0.1
+    - click==8.1.3
+    - cloudpickle==2.2.1
+    - constantly==15.1.0
+    - contourpy==1.0.7
+    - cryptography==39.0.1
+    - cycler==0.11.0
+    - dill==0.3.6
+    - distlib==0.3.6
+    - dm-tree==0.1.8
+    - docker-pycreds==0.4.0
+    - filelock==3.9.0
+    - fire==0.4.0
+    - fonttools==4.38.0
+    - frozenlist==1.3.3
+    - gitdb==4.0.10
+    - gitpython==3.1.31
+    - google-auth==2.16.1
+    - google-auth-oauthlib==0.4.6
+    - gym==0.23.0
+    - gym-notices==0.0.8
+    - gymnasium==0.26.3
+    - gymnasium-notices==0.0.1
+    - h5py==3.7.0
+    - hyperlink==21.0.0
+    - idna==3.4
+    - imageio==2.23.0
+    - importlib-metadata==4.13.0
+    - importlib-resources==5.12.0
+    - incremental==22.10.0
+    - iniconfig==2.0.0
+    - isort==5.12.0
+    - jsonschema==4.17.3
+    - kiwisolver==1.4.4
+    - lazy-loader==0.1
+    - lazy-object-proxy==1.9.0
+    - lz4==4.0.0
+    - markdown==3.4.1
+    - markdown-it-py==2.2.0
+    - markupsafe==2.1.2
+    - matplotlib==3.7.0
+    - mccabe==0.7.0
+    - mdurl==0.1.2
+    - msgpack==1.0.4
+    - networkx==3.0
+    - numpy==1.23.3
+    - oauthlib==3.2.2
+    - opencv-python==3.4.17.63
+    - openskill==2.4.0
+    - ordered-set==4.1.0
+    - packaging==23.0
+    - pandas==1.5.3
+    - pathtools==0.1.2
+    - pettingzoo==1.19.0
+    - pillow==9.4.0
+    - platformdirs==3.0.0
+    - pluggy==1.0.0
+    - protobuf==3.20.3
+    - psutil==5.9.4
+    - py==1.11.0
+    - py-cpuinfo==9.0.0
+    - pyasn1==0.4.8
+    - pyasn1-modules==0.2.8
+    - pycparser==2.21
+    - pyflakes==3.0.1
+    - pygments==2.14.0
+    - pyhamcrest==2.0.4
+    - pylint==2.16.0
+    - pyparsing==3.0.9
+    - pyrsistent==0.19.3
+    - pytest==6.2.5
+    - pytest-benchmark==3.4.1
+    - pytest-pythonpath==0.7.4
+    - python-dateutil==2.8.2
+    - pytz==2022.7.1
+    - pywavelets==1.4.1
+    - pyyaml==6.0
+    - ray==2.3.0
+    - requests==2.28.2
+    - requests-oauthlib==1.3.1
+    - rich==13.3.1
+    - rsa==4.9
+    - scikit-image==0.20.0
+    - scipy==1.9.1
+    - sentry-sdk==1.16.0
+    - service-identity==21.1.0
+    - setproctitle==1.1.10
+    - six==1.16.0
+    - smmap==5.0.0
+    - tabulate==0.9.0
+    - tensorboard==2.12.0
+    - tensorboard-data-server==0.7.0
+    - tensorboard-plugin-wit==1.8.1
+    - tensorboardx==2.6
+    - termcolor==2.2.0
+    - tifffile==2023.2.27
+    - toml==0.10.2
+    - tomli==2.0.1
+    - tomlkit==0.11.6
+    - torch==1.13.1
+    - tqdm==4.61.1
+    - twisted==19.2.0
+    - txaio==23.1.1
+    - typer==0.7.0
+    - typing-extensions==4.5.0
+    - urllib3==1.26.14
+    - vec-noise==1.1.4
+    - virtualenv==20.20.0
+    - wandb==0.13.10
+    - werkzeug==2.2.3
+    - wrapt==1.15.0
+    - zipp==3.15.0
+    - zope-interface==5.5.2
+prefix: /Users/daveey/miniforge3/envs/nmmo
