--- scripts/train_simple.sh
@@ -1,15 +1,0 @@
-#!/bin/bash
-
-python -u -O \
-  -m main \
-  --model.arch=simple \
-  --env.num_teams=8 \
-  --env.team_size=1 \
-  --rollout.num_envs=1 \
-  --rollout.num_buffers=1  \
-  --rollout.num_steps=128 \
-  --wandb.entity=daveey \
-  --wandb.project=nmmo \
-  --train.experiments_dir=/fsx/home-daveey/experiments \
-  --train.num_steps=100000000 \
-  "${@}"
